fixed incorrect return type for sum()

+++ sums.h
@@ -2,6 +2,6 @@
 #define __sum_h_
 
 extern void setup(int64_t N, float A[]);
-extern double sum(int64_t N, float A[]);
+extern float sum(int64_t N, float A[]);
 
 #endif
